+++ jdaviz/configs/imviz/plugins/catalogs/catalogs.py
@@ -68,7 +68,6 @@ class Catalogs(PluginTemplateMixin, ViewerSelectMixin, HasFileImportSelect, Tabl
         return PluginUserApi(self, expose=('clear_table', 'export_table', 'import_catalog',
                                            'zoom_to_selected', 'select_rows',
                                            'select_all', 'select_none',
-                                           'results_available', 'number_of_results',
                                            'catalog', 'max_sources', 'search',
                                            'table', 'table_selected'))